requirements.txt
pysocks
mechanize
fuzzywuzzy[speedup]
beautifulsoup4
